+++ app/src/processing/app/Base.java
@@ -23,27 +23,28 @@
 
 package processing.app;
 
-import java.awt.*;
-import java.awt.event.ActionListener;
-import java.io.*;
-import java.lang.reflect.InvocationTargetException;
-import java.util.*;
-import java.util.List;
-import java.util.Map.Entry;
-
-import javax.swing.*;
-import javax.swing.tree.DefaultMutableTreeNode;
-
 import com.formdev.flatlaf.FlatDarkLaf;
 import com.formdev.flatlaf.FlatLaf;
 import com.formdev.flatlaf.FlatLightLaf;
 import processing.app.contrib.*;
 import processing.app.tools.Tool;
 import processing.app.ui.*;
-import processing.app.ui.PreferencesKt;
 import processing.app.ui.Toolkit;
-import processing.core.*;
+import processing.core.PApplet;
 import processing.data.StringList;
+
+import javax.swing.*;
+import javax.swing.tree.DefaultMutableTreeNode;
+import java.awt.*;
+import java.awt.event.ActionListener;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.lang.reflect.InvocationTargetException;
+import java.util.*;
+import java.util.List;
+import java.util.Map.Entry;
 
 /**
  * The base class for the main processing application.
@@ -2185,11 +2186,10 @@ public class Base {
    * Show the Preferences window.
    */
   public void handlePrefs() {
-//    if (preferencesFrame == null) {
-//      preferencesFrame = new PreferencesFrame(this);
-//    }
-//    preferencesFrame.showFrame();
-      PreferencesKt.show();
+      if (preferencesFrame == null) {
+          preferencesFrame = new PreferencesFrame(this);
+      }
+      preferencesFrame.showFrame();
   }
