+++ app/src/processing/app/ui/PDEWelcome.kt
@@ -6,64 +6,19 @@ import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.tween
 import androidx.compose.animation.slideIn
 import androidx.compose.animation.slideOut
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.VerticalScrollbar
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.defaultMinSize
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.sizeIn
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.rememberScrollbarAdapter
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.outlined.NoteAdd
 import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material.icons.filled.Language
-import androidx.compose.material.icons.outlined.Book
-import androidx.compose.material.icons.outlined.ChatBubbleOutline
-import androidx.compose.material.icons.outlined.FolderOpen
-import androidx.compose.material.icons.outlined.FolderSpecial
-import androidx.compose.material.icons.outlined.PinDrop
-import androidx.compose.material.icons.outlined.School
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.CheckboxDefaults
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.ProvideTextStyle
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateListOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
@@ -79,19 +34,10 @@ import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.application
-import processing.app.Base
-import processing.app.LocalPreferences
-import processing.app.Messages
-import processing.app.Platform
+import processing.app.*
 import processing.app.api.Contributions.ExamplesList.Companion.listAllExamples
 import processing.app.api.Sketch.Companion.Sketch
-import processing.app.ui.preferences.Interface.Companion.languagesDropdown
-import processing.app.ui.theme.LocalLocale
-import processing.app.ui.theme.Locale
-import processing.app.ui.theme.PDEComposeWindow
-import processing.app.ui.theme.PDESwingWindow
-import processing.app.ui.theme.PDETheme
-import processing.app.ui.theme.toDimension
+import processing.app.ui.theme.*
 import java.io.File
 import kotlin.io.path.Path
 import kotlin.io.path.exists
@@ -626,6 +572,27 @@ fun showWelcomeScreen(base: Base? = null) {
     }
 }
 
+@Composable
+fun languagesDropdown(showOptions: MutableState<Boolean>) {
+    val locale = LocalLocale.current
+    val languages = if (Preferences.isInitialized()) Language.getLanguages() else mapOf("en" to "English")
+    DropdownMenu(
+        expanded = showOptions.value,
+        onDismissRequest = {
+            showOptions.value = false
+        },
+    ) {
+        languages.forEach { family ->
+            DropdownMenuItem(
+                text = { Text(family.value) },
+                onClick = {
+                    locale.set(java.util.Locale(family.key))
+                    showOptions.value = false
+                }
+            )
+        }
+    }
+}
 
 fun main(){
     application {
